Reworked structure for code sharing
