+++ packages/preactement/src/parse.ts
@@ -1,30 +1,6 @@
 import { h, ComponentFactory, Fragment } from 'preact';
-import { ErrorTypes, CustomElement, IProps } from './model';
-
-/* -----------------------------------
- *
- * parseJson
- *
- * -------------------------------- */
-
-function parseJson(this: CustomElement, value: string) {
-  const { tagName } = this;
-  const { formatProps } = this.__options;
-
-  let result = {};
-
-  try {
-    result = JSON.parse(value);
-  } catch {
-    console.error(ErrorTypes.Json, `: <${tagName.toLowerCase()}>`);
-  }
-
-  if (formatProps) {
-    result = formatProps(result);
-  }
-
-  return result;
-}
+import { getDocument } from '@/shared/parse';
+import { CustomElement, IProps } from './model';
 
 /* -----------------------------------
  *
@@ -42,30 +18,6 @@ function parseHtml(this: CustomElement): ComponentFactory<{}> {
   const result = convertToVDom.call(this, dom);
 
   return () => result;
-}
-
-/* -----------------------------------
- *
- * getDocument
- *
- * -------------------------------- */
-
-function getDocument(html: string) {
-  const value = `<!DOCTYPE html>\n<html><body>${html}</body></html>`;
-
-  let nodes: Document;
-
-  try {
-    nodes = new DOMParser().parseFromString(value, 'text/html');
-  } catch {
-    // no-op
-  }
-
-  if (!nodes) {
-    return void 0;
-  }
-
-  return nodes.body;
 }
 
 /* -----------------------------------
@@ -182,4 +134,4 @@ function getSlotChildren(children: JSX.Element[]) {
  *
  * -------------------------------- */
 
-export { parseJson, parseHtml, getPropKey, getAttributeProps };
+export { parseHtml, getPropKey, getAttributeProps };
