+++ tsconfig.json
@@ -14,7 +14,7 @@
     "sourceMap": true,
     "lib": ["DOM", "ES2015", "ES2016", "ES2017", "ES2018", "ES2019"],
     "paths": {
-      "@/*": ["./src/*"]
+      "@/*": ["./packages/*"]
     }
   },
   "exclude": ["node_modules", "webpack.config.ts", "tests", "./dist"]
